refactor(Pairs): 단수형으로 클래스 이름 변경
